requirements.txt
requests
beautifulsoup4
tldextract
graphql-core
pyyaml
tqdm